RUB-5170 Updated to use longs for transaction tags. Remove batch query/close and tagged update as they are no longer supported.

+++ src/com/exact/ews/transaction/JsonCoder.java
@@ -148,7 +148,7 @@ class JsonCoder implements Coder {
     if(!cc_number.equals("null"))
       r.setCardNumber(cc_number);
 
-    r.setTransactionTag(json.optInt("transaction_tag"));
+    r.setTransactionTag(json.optLong("transaction_tag"));
     r.setTrack1(json.optString("track1"));
     r.setTrack2(json.optString("track2"));
     r.setPrimaryAccountNumber(json.optString("pan"));
